ModTool: Color Enhancement
* Add a separate ModToolException for the CLI
* Move the CLI Logger code to CLI base module
* Add docstrings to CLI functions
* Fix lint errors
