+++ gr-utils/python/modtool/cli/newmod.py
@@ -26,8 +26,8 @@ import os
 import click
 
 from gnuradio import gr
-from ..core import ModToolException
 from ..core import ModToolNewModule
+from .base import ModToolException
 from .base import common_params, run
 
 
@@ -61,9 +61,10 @@ def cli(**kwargs):
     run(self)
 
 def get_modname(self):
+    """ Get the name of the new module to be added """
     if self.info['modname'] is None:
         while not self.info['modname'] or self.info['modname'].isspace():
-            self.info['modname'] = input(click.style(
-                                         'Name of the new module: ', fg='blue'))
+            self.info['modname'] = input(click.style('Name of the new module: ',
+                                                     fg='blue'))
     if not re.match('[a-zA-Z0-9_]+$', self.info['modname']):
         raise ModToolException('Invalid module name.')
