+++ gr-utils/python/modtool/cli/rm.py
@@ -40,11 +40,11 @@ def cli(**kwargs):
     run(self)
 
 def get_pattern(self):
+    """ Returns the regex pattern for block(s) to be removed """
     if self.info['pattern'] is None:
         block_candidates = get_block_candidates()
         with SequenceCompleter(block_candidates):
-            self.info['pattern'] = input(click.style(
-                                         'Which blocks do you want to delete? (Regex): ',
-                                         fg='blue'))
+            self.info['pattern'] = input(click.style('Which blocks do you want to delete? (Regex): ',
+                                                     fg='blue'))
     if not self.info['pattern'] or self.info['pattern'].isspace():
         self.info['pattern'] = '.'
